python: Refactor to be compatible with Python 2
To be compatible with Python 2, type hints are now placed interface
files. There was also several new features such as `singledispatch`,
`inspect.signature` and some new syntax for expanding lists.

One of the big challenges is making the string API compatible with
both Python 2 (byte) and Python 3 (unicode) strings. The Jsonnet
API will always return unicode strings, but because byte strings are
in such common use in Python 2 we allow them to be used there.